Move static task-firebase methods to FirestoreTask + discard Task::hasDueDate
