+++ app/src/main/java/com/github/houseorganizer/houseorganizer/Task.java
@@ -112,10 +112,6 @@ public class Task {
         return new ArrayList<>(subtasks);
     }
 
-    /* public boolean hasDueDate() {
-        return dueDate != LocalDateTime.MIN;
-    }*/
-
     public LocalDateTime getDueDate() {
         return dueDate;
     }
